점수의 오버플로우 방지, 정수 형변환 시 오차 방지

+++ Assets/Scripts/PlayScene/BulletParticle.cs
@@ -29,7 +29,7 @@ public class BulletParticle : MonoBehaviour
         if (enemy)
         {
             // 일반 피해량
-            int damage = Player.Instance.damageOnNomal;
+            int damage = (int)Player.Instance.damageOnNomal;
             if (type == BulletType.skill)
             {
                 // 스킬 발동 중 피해량
